pyproject.toml
[project]
name = "awaas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diskcache>=5.6.3",
    "geopandas>=1.1.3",
    "langchain>=1.2.16",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.10",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "reportlab>=4.5.0",
    "requests>=2.33.1",
    "streamlit>=1.57.0",
    "supabase>=2.29.0",
    "tenacity>=9.1.4",
]
